feat: enhance ios error
On iOS, when simplePrompt() is called, it can show two pop-ups, the
first one from the system asking the user to allow biometry for the app
and then the one to input biometry. The user can cancel on both
scenarios and errors being thrown should be differente so that the app
can treat them separetly.